pyproject.toml
[tool.poetry]
name = "nautobot-ssot-device42"
version = "0.12.0"
description = "Nautobot plugin for syncing to Device42"
authors = ["Justin Drew <justin.drew@networktocode.com>"]
readme = "README.md"
homepage = "https://gitlab.corp.zmi.global/networking/nautobot-plugin-device42-sync.git"
repository = "https://gitlab.corp.zmi.global/networking/nautobot-plugin-device42-sync.git"
keywords = ["nautobot", "nautobot-plugin"]
include = [
    "README.md",
]
packages = [
    { include = "nautobot_ssot_device42" },
]

[tool.poetry.dependencies]
# Used for local development
nautobot = { version = "^1.1.0" }
python = "^3.7"
diffsync = "^1.3.0"
requests = "^2.26.0"
nautobot-ssot = "^1.0.0"
dnspython = "^2.1.0"
# netutils = "^0.2.2"
netutils = { git = "https://github.com/networktocode/netutils.git", branch = "develop" }

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coverage = "*"
django-debug-toolbar = "*"
flake8 = "*"
invoke = "*"
mkdocs = "*"
pydocstyle = "*"
pylint = "*"
pylint-django = "*"
yamllint = "*"
python-semantic-release = "^7.19.2"

[tool.poetry.extras]
nautobot = ["nautobot"]

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"

[tool.semantic_release]
version_variable = [
    "nautobot_ssot_device42/__init__.py:__version__"
]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
major_on_zero = true
branch = "main"
upload_to_PyPI = false
upload_to_release = false
build_command = "pip install poetry && poetry build"
changelog_file = "CHANGELOG.md"
hvcs = "Gitlab"
hvcs_domain = "gitlab.corp.zmi.global"
